feat: 충전소 간단 정보 모달 임의 디자인 부여
[#338]
